Refactor TTY functions
