+++ os.c
@@ -12,7 +12,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include "tty.h"
+#include "terminal/tty.h"
 #include "module_loader/multiboot.h"
 
 static char *stack[16*1024] __attribute__((section(".bss"))); // 16-bit stack
